Rename SLA to SLO
SLO stands for Service Level Objective
