+++ configs/schema.yml
@@ -50,7 +50,7 @@ definitions:
       algorithm:
         type: string
         enum:
-          - sla
+          - slo
       args:
         schema:
           $ref: "#/definitions/args"
@@ -63,7 +63,7 @@ definitions:
       - capacity_ratio
       - max_replicas
       - min_replicas
-      - sla_seconds
+      - slo_seconds
     properties:
       avs_task_duration:
         type: integer
@@ -73,5 +73,5 @@ definitions:
         type: integer
       min_replicas:
         type: integer
-      sla_seconds:
+      slo_seconds:
         type: integer
